Refactor rule to check entire rules instead of individual declarations
