+++ lib/rules/index.cjs
@@ -104,10 +104,8 @@ const rules = {
 	get 'declaration-block-no-duplicate-properties'() {
 		return Promise.resolve().then(() => /*#__PURE__*/_interopNamespaceDefaultOnly(require('./declaration-block-no-duplicate-properties/index.cjs'))).then((m) => m.default);
 	},
-	get 'declaration-block-nesting-at-rule-required-list'() {
-		return Promise.resolve().then(() => /*#__PURE__*/_interopNamespaceDefaultOnly(require('./declaration-block-nesting-at-rule-required-list/index.cjs'))).then(
-			(m) => m.default,
-		);
+	get 'rule-nesting-at-rule-required-list'() {
+		return Promise.resolve().then(() => /*#__PURE__*/_interopNamespaceDefaultOnly(require('./rule-nesting-at-rule-required-list/index.cjs'))).then((m) => m.default);
 	},
 	get 'declaration-block-no-redundant-longhand-properties'() {
 		return Promise.resolve().then(() => /*#__PURE__*/_interopNamespaceDefaultOnly(require('./declaration-block-no-redundant-longhand-properties/index.cjs'))).then(
