+++ lib/rules/index.mjs
@@ -98,10 +98,8 @@ const rules = {
 	get 'declaration-block-no-duplicate-properties'() {
 		return import('./declaration-block-no-duplicate-properties/index.mjs').then((m) => m.default);
 	},
-	get 'declaration-block-nesting-at-rule-required-list'() {
-		return import('./declaration-block-nesting-at-rule-required-list/index.mjs').then(
-			(m) => m.default,
-		);
+	get 'rule-nesting-at-rule-required-list'() {
+		return import('./rule-nesting-at-rule-required-list/index.mjs').then((m) => m.default);
 	},
 	get 'declaration-block-no-redundant-longhand-properties'() {
 		return import('./declaration-block-no-redundant-longhand-properties/index.mjs').then(
