+++ lib/rules/rule-nesting-at-rule-required-list/index.cjs
@@ -3,26 +3,25 @@
 'use strict';
 
 const validateTypes = require('../../utils/validateTypes.cjs');
-const eachDeclarationBlock = require('../../utils/eachDeclarationBlock.cjs');
-const isStandardSyntaxDeclaration = require('../../utils/isStandardSyntaxDeclaration.cjs');
 const matchesStringOrRegExp = require('../../utils/matchesStringOrRegExp.cjs');
 const report = require('../../utils/report.cjs');
 const ruleMessages = require('../../utils/ruleMessages.cjs');
 const validateOptions = require('../../utils/validateOptions.cjs');
 
-const ruleName = 'declaration-block-nesting-at-rule-required-list';
+const ruleName = 'rule-nesting-at-rule-required-list';
 
 const messages = ruleMessages(ruleName, {
-	expected: (atRule) => `Expected nesting at-rule "${atRule}" for declaration`,
+	expected: (atRule) => `Expected nesting at-rule "${atRule}" for rule`,
 });
 
 const meta = {
-	url: 'https://stylelint.io/user-guide/rules/declaration-block-nesting-at-rule-required-list',
+	url: 'https://stylelint.io/user-guide/rules/rule-nesting-at-rule-required-list',
 };
 
 /** @type {import('stylelint').CoreRules[ruleName]} */
 const rule = (primary) => {
 	return (root, result) => {
+		// Validate that the primary option is a string, regex, or array of strings/regexes
 		const validOptions = validateOptions(result, ruleName, {
 			actual: primary,
 			possible: [validateTypes.isString, validateTypes.isRegExp],
@@ -30,34 +29,34 @@ const rule = (primary) => {
 
 		if (!validOptions) return;
 
-		eachDeclarationBlock(root, (eachDecl) => {
-			eachDecl((decl) => {
-				if (!isStandardSyntaxDeclaration(decl)) return;
+		root.walkRules((ruleNode) => {
+			// Skip if rule already has a required at-rule ancestor
+			if (hasRequiredAtRuleAncestor(ruleNode, primary)) return;
 
-				// Skip if declaration already has a required at-rule ancestor
-				if (hasRequiredAtRuleAncestor(decl, primary)) return;
-
-				report({
-					message: messages.expected,
-					messageArgs: [Array.isArray(primary) ? primary.join(', ') : primary],
-					node: decl,
-					result,
-					ruleName,
-				});
+			report({
+				message: messages.expected,
+				messageArgs: [Array.isArray(primary) ? primary.join(', ') : primary],
+				node: ruleNode,
+				result,
+				ruleName,
 			});
 		});
 	};
 };
 
 /**
- * @param {import('postcss').Declaration} decl
- * @param {string | RegExp | Array<string | RegExp>} requiredAtRules
- * @returns {boolean}
+ * Check if rule has a required at-rule ancestor
+ * Traverses up the PostCSS AST tree from the rule to find
+ * if any parent node is an at-rule that matches the required patterns
+ * @param {import('postcss').Rule} ruleNode - The CSS rule to check
+ * @param {string | RegExp | Array<string | RegExp>} requiredAtRules - Patterns to match at-rule names against
+ * @returns {boolean} True if rule has required at-rule ancestor, false otherwise
  */
-function hasRequiredAtRuleAncestor(decl, requiredAtRules) {
-	let current = decl.parent;
+function hasRequiredAtRuleAncestor(ruleNode, requiredAtRules) {
+	let current = ruleNode.parent;
 
 	while (current) {
+		// Check if current node is an at-rule (e.g., @layer, @scope, @supports)
 		if (current.type === 'atrule') {
 			const atRuleName = current.name;
